(WIP) Minor changes
